requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyJWT==2.7.0
Werkzeug==2.3.6
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7